requirements.txt
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
dill
pyyaml
ipykernel
xgboost

#-e .